feat(front): add HeaderToggle

+++ client/src/components/Header.jsx
@@ -1,34 +1,29 @@
-import { Link, useLocation, useNavigate } from "react-router-dom";
+import { Link, useLocation } from "react-router-dom";
+import { usePage } from "../contexts/PageContext";
+import { useState, useEffect } from "react";
+import { ConfirmationModal } from "./ConfirmationModal";
 import logo from "../assets/img/Logo.png";
 import trash from "../assets/img/Trash.png";
 import { IoIosArrowBack } from "react-icons/io";
-import { usePage } from "../contexts/PageContext";
-import { useState } from "react";
-import { ConfirmationModal } from "./ConfirmationModal";
-import { useEffect } from "react";
+import { HeaderToggle } from "./HeaderToggle";
 
 export const Header = () => {
   const [isModalOpen, setIsModalOpen] = useState(false);
   const location = useLocation();
-  const navigate = useNavigate();
-  const isRoot = location.pathname === "/";
   const { pageData, setPageData } = usePage();
 
+  const isRoot = location.pathname === "/alunos" || location.pathname === "/cursos";
+
   const handleDelete = async () => {
-    if (pageData.onDelete) {
-      await pageData.onDelete();
-      navigate("/");
-    }
+    if (pageData.onDelete) await pageData.onDelete();
     setIsModalOpen(false);
   };
 
   useEffect(() => {
-    if (location.pathname === "/") {
-      setPageData({
-        title: "Gerenciador de alunos",
-        subtitle: "",
-        onDelete: null,
-      });
+    if (location.pathname === "/alunos") {
+      setPageData({ title: "Alunos", onDelete: null });
+    } else if (location.pathname === "/cursos") {
+      setPageData({ title: "Cursos", onDelete: null });
     }
   }, [location.pathname, setPageData]);
 
@@ -41,31 +36,29 @@ export const Header = () => {
               <IoIosArrowBack className="text-white text-2xl" />
             </Link>
           )}
-          {isRoot && (
-            <Link to="/">
-              <img src={logo} alt="Logo" />
-            </Link>
-          )}
+          <Link to={isRoot ? location.pathname : "/"}>
+            <img src={logo} alt="Logo" />
+          </Link>
           <h1 className="text-white font-extrabold tracking-[1%] leading-[30px] flex items-center gap-2">
             {pageData.title}
-            {pageData.subtitle && (
-              <span className="font-medium text-white text-sm">
-                | {pageData.subtitle}
-              </span>
-            )}
           </h1>
         </div>
 
-        {!isRoot && pageData.onDelete && (
-          <button
-            onClick={() => setIsModalOpen(true)}
-            className="cursor-pointer w-5 h-5"
-            title="Excluir"
-          >
-            <img className="w-full" src={trash} alt="Trash icon" />
-          </button>
+        {isRoot ? (
+          <HeaderToggle />
+        ) : (
+          pageData.onDelete && (
+            <button
+              onClick={() => setIsModalOpen(true)}
+              className="cursor-pointer w-5 h-5"
+              title="Excluir"
+            >
+              <img className="w-full" src={trash} alt="Trash icon" />
+            </button>
+          )
         )}
       </div>
+
       <ConfirmationModal
         isOpen={isModalOpen}
         onClose={() => setIsModalOpen(false)}
